% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/working_functions.R
\name{empty_columns}
\alias{empty_columns}
\title{Detect empty columns}
\usage{
empty_columns(.data)
}
\arguments{
\item{.data}{Character vector, data.frame or a list.}
}
\value{
a logical, indicating empty data.
}
\description{
Detect empty columns
}
