% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/working_functions.R
\name{fec}
\alias{fec}
\title{Filter empty columns}
\usage{
fec(.data)
}
\arguments{
\item{.data}{Character vector, data frame or list to filter in.}
}
\value{
An object of the same type of \code{.data}, but without elements that are all
missing
}
\description{
Filter  empty columns
}
\examples{
tracking_vector <- c(NA, 0)
fec(tracking_vector)
data(mtcars)
mtcars$newempty <- NA
fec(mtcars)
mtcars$another_new <- sample(c(NA, ""), nrow(mtcars), replace = TRUE)
fec(mtcars)
}
