% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/working_functions.R
\name{filter_empty_columns}
\alias{filter_empty_columns}
\title{Filtering empty columns}
\usage{
filter_empty_columns(.data)
}
\arguments{
\item{.data}{Character vector, data frame or list to filter in.}
}
\value{
An object of the same type of \code{.data}, but without elements that are all
missing
}
\description{
Filtering on empty columns
}
\examples{
tracking_vector <- c(NA, 0)
filter_empty_columns(tracking_vector)
data(mtcars)
mtcars$newempty <- NA
filter_empty_columns(mtcars)
mtcars$another_new <- sample(c(NA, ""), nrow(mtcars), replace = TRUE)
filter_empty_columns(mtcars)
}
