% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertAssignPipeAddin.R
\name{insertAssignPipeAddin}
\alias{insertAssignPipeAddin}
\title{Insert \%<>\%.}
\usage{
insertAssignPipeAddin()
}
\description{
Call this function as an addin to insert \code{ \%<>\% } at the cursor position.
}

