% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_functions.R
\name{InsertCompound}
\alias{InsertCompound}
\title{Insert the magrittr compound assignment pipe-operator}
\usage{
InsertCompound()
}
\description{
Insert the magrittr compound assignment pipe-operator
}
