% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_functions.R
\name{InsertCustomNames}
\alias{InsertCustomNames}
\title{Insert the cnr (custom name repair) function when loading a data frame from tibble}
\usage{
InsertCustomNames()
}
\description{
Insert the cnr (custom name repair) function when loading a data frame from tibble
}
