% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_functions.R
\name{InsertFiltering}
\alias{InsertFiltering}
\title{Insert the filter empty column and return back function}
\usage{
InsertFiltering()
}
\description{
Insert the filter empty column and return back function
}
