% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_functions.R
\name{InsertGeq}
\alias{InsertGeq}
\title{Insert Greather or equal than (ideal on keyboards without access)}
\usage{
InsertGeq()
}
\description{
Insert Greather or equal than (ideal on keyboards without access)
}
