% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_functions.R
\name{InsertLeq}
\alias{InsertLeq}
\title{Insert less or equal than (ideal on keyboard without access)}
\usage{
InsertLeq()
}
\description{
Insert less or equal than (ideal on keyboard without access)
}
